Remove automated tests
With some help, I realized that the test setup is not going to work for
our environment.  Rather than trying to troublshoot the tests, I decided
to skip having tests for the role.  The alternative is to modify the
base test environment to match the behavior & configuration of the RHEL
hosts that are provided to us by OIT.  That felt like a lot of work, and
potentially trying to hit a moving target.